requirements.txt
appdirs==1.4.3
Appium-Python-Client==0.24
enum34==1.1.6
glob2==0.5
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
parse==1.8.0
parse-type==0.3.4
py==1.4.33
pyparsing==2.2.0
pytest==3.0.7
pytest-bdd==2.18.1
selenium==3.4.0
six==1.10.0
